log_buf_addr_get() and log_buf_len_get() not available to external modules
They are defined by CONFIG_PRINTK in include/linux/printk.h,
and implemented in kernel/printk/printk.c, but not exported.

They are visible in /proc/kallsyms when CONFIG_PRINTK is on, but
not available to external kernel modules. Only for built-ins.

There is no BUILT-IN define unfortunately; so !MODULE would do.

Hopefully they are not used!